requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.10
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
numpy==1.20.2
pandas==1.2.3
Pillow==8.2.0
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.20.1
scipy==1.6.3
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.26.4
waitress==2.0.0
Werkzeug==1.0.1
yacs==0.1.8